+++ src/language-js/sortSwitchCases/index.ts
@@ -1,4 +1,9 @@
-import { BaseExpression, Comment, SwitchCase } from "estree";
+import {
+  BaseExpression,
+  BaseNodeWithoutComments,
+  Comment,
+  SwitchCase
+} from "estree";
 
 import {
   BaseNode,
@@ -49,24 +54,7 @@ export function sortSwitchCases(
     let switchCaseEnd = 0;
     for (switchCaseEnd = 0; switchCaseEnd < cases.length; switchCaseEnd++) {
       // Do not rearrange items that are in a non-break statement
-      let breakStatement = cases[switchCaseEnd].consequent.filter(
-        (value: any) => {
-          if (value.type === "BlockStatement") {
-            return (
-              value.body.filter((value: any) => {
-                return (
-                  value.type === "BreakStatement" ||
-                  value.type === "ReturnStatement"
-                );
-              }).length !== 0
-            );
-          }
-          return (
-            value.type === "BreakStatement" || value.type === "ReturnStatement"
-          );
-        }
-      );
-      if (breakStatement.length === 0) {
+      if (!doesCaseBreakOutOfSwitch(cases[switchCaseEnd])) {
         continue;
       }
 
@@ -120,6 +108,14 @@ export function sortSwitchCases(
       }
     }
 
+    // If the last switch group is a fall through, don't include it in the swap
+    if (!doesCaseBreakOutOfSwitch(cases[cases.length - 1])) {
+      switchGroupsWithBreaks = switchGroupsWithBreaks.slice(
+        0,
+        switchGroupsWithBreaks.length - 1
+      );
+    }
+
     // Now sort the actual switch groups
     let switchGroupsWithBreaksSorted = switchGroupsWithBreaks.slice();
     let switchGroupToLowestCase = new Map();
@@ -169,6 +165,22 @@ export function sortSwitchCases(
   return newFileContents;
 }
 
+function doesCaseBreakOutOfSwitch(caseStatement: any) {
+  let breakStatement = caseStatement.consequent.filter((value: any) => {
+    if (value.type === "BlockStatement") {
+      return (
+        value.body.filter((value: any) => {
+          return (
+            value.type === "BreakStatement" || value.type === "ReturnStatement"
+          );
+        }).length !== 0
+      );
+    }
+    return value.type === "BreakStatement" || value.type === "ReturnStatement";
+  });
+  return breakStatement.length !== 0;
+}
+
 function getSortableText(a: any, fileContents: string) {
   if (a == null) {
     return null;
@@ -189,7 +201,9 @@ function caseGroupsToMinimumTypeinformations(
   switchGroupsWithBreaks: SwitchCase[][]
 ) {
   return switchGroupsWithBreaks.map(value => {
-    let firstRange = value[0].range;
+    let firstNode: BaseNodeWithoutComments = value[0];
+
+    let firstRange = firstNode.range;
     let lastRange = value[value.length - 1].range;
     if (firstRange == null || lastRange == null) {
       throw new Error("Range is null");