+++ src/utilities/sort-utils.test.ts
@@ -12,8 +12,29 @@ describe("utilities/sort-utils", () => {
     getContextGroups(nodes, comments, "");
   });
 
-  describe("Single context group", () => {
-    it("comment", () => {
+  it("Spaces within nodes but not around", () => {
+    let input = `
+    {
+      console.log("a");
+      
+      console.log("b");
+    }
+    {
+      console.log("c");
+      
+      console.log("d");
+    }
+    `;
+    let parsed = flowParse(input);
+    let result = getContextGroups(parsed.body, parsed.comments, input);
+
+    expect(result).to.have.lengthOf(1);
+    expect(result[0].nodes).to.have.lengthOf(2);
+    expect(result[0].comments).to.have.lengthOf(0);
+  });
+
+  describe("Comments - Single context group", () => {
+    it("comment top", () => {
       let input = `
       // First
       `;
@@ -37,7 +58,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
     });
 
-    it("comment, node", () => {
+    it("comment top, node", () => {
       let input = `
       // First
       log("First");
@@ -50,7 +71,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
     });
 
-    it("comment, comment, node", () => {
+    it("comment top, comment top, node", () => {
       let input = `
       // First
       // First - 2
@@ -64,7 +85,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
     });
 
-    it("comment, node, comment, node", () => {
+    it("comment top, node, comment top, node", () => {
       let input = `
       // First
       log("First");
@@ -79,7 +100,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(2);
     });
 
-    it("comment, node, node", () => {
+    it("comment top, node, node", () => {
       let input = `
       // First
       log("First");
@@ -93,7 +114,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
     });
 
-    it("node, comment", () => {
+    it("node, comment top", () => {
       let input = `
       log("First");
       // Second
@@ -106,7 +127,7 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
     });
 
-    it("node, comment, node", () => {
+    it("node, comment top, node", () => {
       let input = `
       log("First");
       // Second
@@ -119,9 +140,98 @@ describe("utilities/sort-utils", () => {
       expect(result[0].nodes).to.have.lengthOf(2);
       expect(result[0].comments).to.have.lengthOf(1);
     });
+
+    it("node, comment right", () => {
+      let input = `
+      log("First")  // First;
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(1);
+      expect(result[0].comments).to.have.lengthOf(1);
+    });
+
+    it("comment, node, comment right", () => {
+      let input = `
+      // First
+      log("First")  // First
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(1);
+      expect(result[0].comments).to.have.lengthOf(2);
+    });
+
+    it("node, comment right, node", () => {
+      let input = `
+      log("First"); // First
+      log("Second");
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(2);
+      expect(result[0].comments).to.have.lengthOf(1);
+    });
+
+    it("node, block comment right, node", () => {
+      let input = `
+      log("First"); /* First */
+      log("Second");
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(2);
+      expect(result[0].comments).to.have.lengthOf(1);
+    });
+
+    it("node, comment right, node, comment right", () => {
+      let input = `
+      log("First"); // First
+      log("Second"); // Second
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(2);
+      expect(result[0].comments).to.have.lengthOf(2);
+    });
+
+    it("comment, node, comment, node all on one line", () => {
+      let input = `
+      /* First */ log("First"); /* Second */ log("Second");
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(2);
+      expect(result[0].comments).to.have.lengthOf(2);
+    });
+
+    it("node, comment right, break, comment front, node, comment right", () => {
+      let input = `
+      log("First"); // First
+      /* Second */ log("Second");
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(2);
+      expect(result[0].comments).to.have.lengthOf(2);
+    });
   });
 
-  describe("Multiple context groups", () => {
+  describe("Comments - Multiple context groups", () => {
     it("comment, node, break, comment, node", () => {
       let input = `
       // First
@@ -175,6 +285,58 @@ describe("utilities/sort-utils", () => {
       expect(result[0].comments).to.have.lengthOf(0);
       expect(result[1].nodes).to.have.lengthOf(1);
       expect(result[1].comments).to.have.lengthOf(0);
+    });
+
+    it("comment, break, comment top, comment front, node, comment right", () => {
+      let input = `
+      // First
+
+      // Second
+      /* Second2 */ log("Second"); // Second-3
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(1);
+      expect(result[0].nodes).to.have.lengthOf(1);
+      expect(result[0].comments).to.have.lengthOf(3);
+    });
+
+    it("comment top, comment front, node, comment right, break, comment top, comment front, node, comment right", () => {
+      let input = `
+      // First
+      /* First2 */ log("First"); // First-3
+
+      // Second
+      /* Second2 */ log("Second"); // Second-3
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(2);
+      expect(result[0].nodes).to.have.lengthOf(1);
+      expect(result[0].comments).to.have.lengthOf(3);
+      expect(result[1].nodes).to.have.lengthOf(1);
+      expect(result[1].comments).to.have.lengthOf(3);
+    });
+
+    it("comment top, comment front, node, comment right, break, comment top, comment top, comment front, node, comment right", () => {
+      let input = `
+      // First
+      /* First2 */ log("First"); // First-3
+
+      // Second1
+      // Second11
+      /* Second2 */ log("Second"); // Second-3
+      `;
+      let parsed = flowParse(input);
+      let result = getContextGroups(parsed.body, parsed.comments, input);
+
+      expect(result).to.have.lengthOf(2);
+      expect(result[0].nodes).to.have.lengthOf(1);
+      expect(result[0].comments).to.have.lengthOf(3);
+      expect(result[1].nodes).to.have.lengthOf(1);
+      expect(result[1].comments).to.have.lengthOf(4);
     });
   });
 });