+++ changelog.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+- We now sort comments to the right of a line not just to the top and to the left
+- Improved handling of switch statements
+
 ### 2.4.0
 
 - Added `css`, `scss` and `less` file support